Use complete source path instead of filename for matching
Especially on the Artifact side that means that the filename is not
"constructed" anymore by the extension. We make no assumption about the
structure of the source file name or path.
The only assumption that is left is, that Maven will resolve to the same
file path, if the same artifact coordinates are resolved multiple times
within a build / session.